+++ Includes/LknIntegrationRedeForWoocommerceWcRedeCreditBlocks.php
@@ -37,6 +37,14 @@ final class LknIntegrationRedeForWoocommerceWcRedeCreditBlocks extends AbstractP
             '1.0.0',
             true
         );
+        wp_localize_script(
+            'rede_credit-blocks-integration',
+            'redeCreditAjax',
+            array(
+                'ajaxurl' => admin_url('admin-ajax.php'),
+                'nonce'   => wp_create_nonce('redeCardNonce'),
+            )
+        );
         if (function_exists('wp_set_script_translations')) {
             wp_set_script_translations('rede_credit-blocks-integration');
         }
@@ -76,10 +84,9 @@ final class LknIntegrationRedeForWoocommerceWcRedeCreditBlocks extends AbstractP
             ($settings['installment_interest'] === 'yes' || $settings['installment_discount']) &&
             is_plugin_active('rede-for-woocommerce-pro/rede-for-woocommerce-pro.php')
         ) {
-
             for ($i = 1; $i <= $maxParcels; ++$i) {
                 $parcelAmount = $cart_total / $i;
-                if ($parcelAmount >= $minParcelValue && isset($settings[$i . 'x'])) {
+                if (($i === 1 || $parcelAmount >= $minParcelValue) && isset($settings[$i . 'x'])) {
                     $interest = round((float) $settings[$i . 'x'], 2);
                     $customLabel = LknIntegrationRedeForWoocommerceHelper::lknIntegrationRedeProRedeInterest($cart_total, $interest, $i, 'label', $this->gateway);
                     if ($customLabel) {
@@ -90,7 +97,7 @@ final class LknIntegrationRedeForWoocommerceWcRedeCreditBlocks extends AbstractP
         } else {
             for ($i = 1; $i <= $maxParcels; ++$i) {
                 $parcelAmount = $cart_total / $i;
-                if ($parcelAmount >= $minParcelValue) {
+                if ($i === 1 || $parcelAmount >= $minParcelValue) {
                     $phpArray[$i . 'x'] = html_entity_decode(sprintf('%dx de %s', $i, wp_strip_all_tags(wc_price($parcelAmount))));
                 }
             }